requirements.txt
# Core dependencies
supabase==2.3.4
python-dotenv==1.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0

# Reddit API
praw==7.7.1

# Social media APIs
tweepy==4.14.0
facebook-sdk==3.1.0

# Text processing and NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
scikit-learn==1.3.2

# LLM APIs
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2

# Data processing
pandas==2.1.4
numpy==1.24.4

# Similarity and deduplication
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Utilities
uuid==1.30
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Optional: Database alternatives
# psycopg2-binary==2.9.9  # For direct PostgreSQL connection
# sqlalchemy==2.0.23      # For ORM if needed